+++ packages/at_onboarding_cli/lib/src/activate_cli/activate_cli.dart
@@ -1,29 +1,35 @@
 import 'package:at_commons/at_commons.dart';
 import 'package:at_onboarding_cli/at_onboarding_cli.dart';
 import 'package:args/args.dart';
+import 'package:at_onboarding_cli/src/util/at_onboarding_exceptions.dart';
 import 'dart:io';
 import 'package:at_utils/at_logger.dart';
 
 Future<void> main(List<String> arguments) async {
   //defaults
   String rootServer = 'root.atsign.org';
-  AtSignLogger.root_level = 'severe';
+  String registrarUrl = 'my.atsign.com';
+  AtSignLogger.root_level = 'finer';
 
   //get atSign and CRAM key from args
   final parser = ArgParser()
     ..addOption('atsign', abbr: 'a', help: 'atSign to activate')
-    ..addOption('cramkey', abbr: 'c', help: 'CRAM key')
-    ..addOption('qr_path', abbr: 'q', help: 'path to qr code')
+    ..addOption('cramkey', abbr: 'c', help: 'CRAM key', mandatory: false)
     ..addOption('rootServer',
         abbr: 'r',
-        help: 'root server',
+        help: 'root server\'s domain name',
         defaultsTo: rootServer,
         mandatory: false)
+    ..addOption('registrarUrl',
+        abbr: 'g',
+        help: 'url to the registrar api',
+        mandatory: false,
+        defaultsTo: registrarUrl)
     ..addFlag('help', abbr: 'h', help: 'Usage instructions', negatable: false);
   ArgResults argResults = parser.parse(arguments);
 
   if (argResults.wasParsed('help')) {
-    print(parser.usage);
+    stdout.writeln(parser.usage);
     return;
   }
 
@@ -33,34 +39,40 @@ Future<void> main(List<String> arguments) async {
     throw IllegalArgumentException('atSign is required');
   }
 
-  if (!argResults.wasParsed('cramkey') && !argResults.wasParsed('qr_path')) {
-    stderr.writeln(
-        'Either of --cram key(-c) or --qr_path(-q) is required. Run with --help (or -h) for more.');
-    throw IllegalArgumentException('CRAM key is required');
-  }
+  await activate(argResults);
+}
 
+Future<void> activate(ArgResults argResults,
+    {AtOnboardingService? atOnboardingService}) async {
   stdout.writeln('[Information] Root server is ${argResults['rootServer']}');
-
+  stdout.writeln(
+      '[Information] Registrar url provided is ${argResults['registrarUrl']}');
   //onboarding preference builder can be used to set onboardingService parameters
   AtOnboardingPreference atOnboardingPreference = AtOnboardingPreference()
     ..rootDomain = argResults['rootServer']
-    ..cramSecret = argResults['cramkey'] ??
-        AtOnboardingServiceImpl.getSecretFromQr(argResults['qr_path']);
-
+    ..registrarUrl = argResults['registrarUrl']
+    ..cramSecret =
+        argResults.wasParsed('cramkey') ? argResults['cramkey'] : null;
   //onboard the atSign
-  AtOnboardingService onboardingService =
+  atOnboardingService ??=
       AtOnboardingServiceImpl(argResults['atsign'], atOnboardingPreference);
   stdout.writeln(
       '[Information] Activating your atSign. This may take up to 2 minutes.');
   try {
-    await onboardingService.onboard();
+    await atOnboardingService.onboard();
+  } on InvalidDataException catch (e) {
+    stderr.writeln(
+        '[Error] Activation failed. Invalid data provided by user. Please try again\nCause: ${e.message}');
+  } on InvalidRequestException catch (e) {
+    stderr.writeln(
+        '[Error] Activation failed. Invalid data provided by user. Please try again\nCause: ${e.message}');
+  } on AtActivateException catch (e) {
+    stdout.writeln('[Error] $e');
   } on Exception catch (e) {
     stderr.writeln(
         '[Error] Activation failed. It looks like something went wrong on our side.\n'
-        'Please try again or contact support@atsign.com\nCause: ${e.toString()}');
-    return;
+        'Please try again or contact support@atsign.com\nCause: $e');
   } finally {
-    await onboardingService.close();
+    await atOnboardingService.close();
   }
-  return;
 }