+++ packages/at_onboarding_cli/lib/src/onboard/at_onboarding_service_impl.dart
@@ -5,17 +5,21 @@ import 'dart:io';
 
 import 'package:at_chops/at_chops.dart';
 import 'package:at_client/at_client.dart';
+import 'package:at_onboarding_cli/src/util/at_onboarding_exceptions.dart';
+import 'package:at_server_status/at_server_status.dart';
 import 'package:at_utils/at_utils.dart';
 import 'package:at_commons/at_builders.dart';
 import 'package:at_onboarding_cli/src/factory/service_factories.dart';
 import 'package:at_lookup/at_lookup.dart';
-import 'package:at_server_status/at_server_status.dart';
 import 'package:at_onboarding_cli/at_onboarding_cli.dart';
 import 'package:crypton/crypton.dart';
 import 'package:encrypt/encrypt.dart';
 import 'package:zxing2/qrcode.dart';
 import 'package:image/image.dart';
 import 'package:path/path.dart' as path;
+
+import '../util/home_directory_util.dart';
+import '../util/onboarding_util.dart';
 
 ///class containing service that can onboard/activate/authenticate @signs
 class AtOnboardingServiceImpl implements AtOnboardingService {
@@ -35,7 +39,7 @@ class AtOnboardingServiceImpl implements AtOnboardingService {
 
   AtOnboardingServiceImpl(atsign, this.atOnboardingPreference,
       {this.atServiceFactory}) {
-    //performs atSign format checks on the atSign
+    // performs atSign format checks on the atSign
     _atSign = AtUtils.fixAtSign(atsign);
 
     // set default LocalStorage paths for this instance
@@ -79,32 +83,31 @@ class AtOnboardingServiceImpl implements AtOnboardingService {
 
   @override
   Future<bool> onboard() async {
-    // TODO: uncomment this code after isOnboarded is implemented
-    // if(isOnboarded()) {
-    //   return true;
-    // }
-    //get cram_secret from either from AtOnboardingConfig or decode it from qr code whichever available
-    atOnboardingPreference.cramSecret ??=
-        getSecretFromQr(atOnboardingPreference.qrCodePath);
-
-    if (atOnboardingPreference.cramSecret == null) {
-      throw AtClientException.message(
-          'Either of cram secret or qr code containing cram secret not provided',
-          exceptionScenario: ExceptionScenario.invalidValueProvided);
-    }
-    if (atOnboardingPreference.downloadPath == null &&
-        atOnboardingPreference.atKeysFilePath == null) {
-      throw AtClientException.message(
-          'Download path not provided. Please provide'
-          ' downloadPath through AtOnboardingPreferences.\n DownloadPath is where the atKeys file will be saved.',
-          exceptionScenario: ExceptionScenario.invalidValueProvided);
-    }
-    // cram auth doesn't use at_chops.So create at_lookup here.
+    // cram auth doesn't use at_chops. So create at_lookup here.
     AtLookupImpl atLookUpImpl = AtLookupImpl(_atSign,
         atOnboardingPreference.rootDomain, atOnboardingPreference.rootPort);
+
+    // get cram_secret from either from AtOnboardingPreference
+    // or fetch from the registrar using verification code sent to email
+    atOnboardingPreference.cramSecret ??= await OnboardingUtil()
+        .getCramUsingOtp(_atSign, atOnboardingPreference.registrarUrl);
+    if (atOnboardingPreference.cramSecret == null) {
+      logger.info('Root Server address is ${atOnboardingPreference.rootDomain}:'
+          '${atOnboardingPreference.rootPort}');
+      logger
+          .info('Registrar url is \'${atOnboardingPreference.registrarUrl}\'');
+      throw AtKeyNotFoundException(
+          'Could not fetch cram secret for \'$_atSign\' from registrar');
+    }
+
+    //check and wait till secondary exists
+    await _waitUntilSecondaryCreated(atLookUpImpl);
+
+    if (await isOnboarded()) {
+      throw AtActivateException('atsign is already activated');
+    }
+
     try {
-      //check and wait till secondary exists
-      await _waitUntilSecondaryCreated(atLookUpImpl);
       //authenticate into secondary using cram secret
       _isAtsignOnboarded = (await atLookUpImpl
           .authenticate_cram(atOnboardingPreference.cramSecret));
@@ -116,9 +119,10 @@ class AtOnboardingServiceImpl implements AtOnboardingService {
         logger.severe('Cram authentication failed. Please check the cram key'
             'and try again \n(or) contact support@atsign.com');
       }
-    } on UnAuthenticatedException {
-      throw UnAuthenticatedException(
-          "Unable to authenticate. Please check your cram secret");
+    } on Exception catch (e) {
+      logger.severe('Caught exception: $e');
+    } on Error catch(e){
+      logger.severe('Caught error: $e');
     } finally {
       await atLookUpImpl.close();
     }
@@ -398,15 +402,36 @@ class AtOnboardingServiceImpl implements AtOnboardingService {
     return atServerStatus.get(_atSign);
   }
 
+  @override
+  Future<bool> isOnboarded() async {
+    late AtStatus secondaryStatus;
+    try {
+      secondaryStatus = await getServerStatus();
+    } catch (e) {
+      stderr.writeln('[Error] $e');
+    }
+    if (secondaryStatus.status() == AtSignStatus.activated) {
+      _isAtsignOnboarded = true;
+      return _isAtsignOnboarded;
+    } else if (secondaryStatus.status() == AtSignStatus.teapot) {
+      return false;
+    }
+    stderr.writeln(
+        '[Error] atsign($_atSign) status is \'${secondaryStatus.status()!.name}\'');
+    throw AtActivateException('Could not determine atsign activation status',
+        intent: Intent.fetchData);
+  }
+
   ///extracts cram secret from qrCode
+  @Deprecated('qr_code based cram authentication not supported anymore')
   static String? getSecretFromQr(String? path) {
     if (path == null) {
       return null;
     }
     try {
       Image? image = decodePng(File(path).readAsBytesSync());
-      LuminanceSource source = RGBLuminanceSource(image!.width, image.height,
-          image.getBytes().buffer.asInt32List());
+      LuminanceSource source = RGBLuminanceSource(
+          image!.width, image.height, image.getBytes().buffer.asInt32List());
       BinaryBitmap bitmap = BinaryBitmap(HybridBinarizer(source));
       Result result = QRCodeReader().decode(bitmap);
       String secret = result.text.split(':')[1];
@@ -478,7 +503,8 @@ class AtOnboardingServiceImpl implements AtOnboardingService {
     }
     _atLookUp = null;
     _atClient = null;
-    logger.info('Closing current instance of at_onboarding_cli');
+    logger.info(
+        'Closing current instance of at_onboarding_cli (exit code: $exitCode)');
   }
 
   @override
@@ -505,11 +531,4 @@ class AtOnboardingServiceImpl implements AtOnboardingService {
 
   @override
   AtChops? atChops;
-
-  @override
-  Future<bool> isOnboarded() async {
-    // #TODO implement once AtClient offline access feature is complete.
-    // https://github.com/atsign-foundation/at_client_sdk/issues/915
-    throw UnimplementedError();
-  }
 }