+++ packages/at_onboarding_cli/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.3.1
+- feat: Introduced verification-code based activation of atsigns
+- fix: deprecate qr_code based activation
+- feat: introduced new exceptions
+- fix: minor bug fixes
 ## 1.2.6
 - feat: changes to integrate onboarding_cli with pkam secure element
 - fix: issue with atKeys file creation while onboarding if the downloadPath directory does not exist